gradleized and upgraded to log4j2
