+++ .gitignore
@@ -1,8 +1,9 @@
-*.iml
-target/
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
 .idea/
-classes/
-*.swagger.json
-*.DS_Store
-metadata-full/
-res/
+*.log
+gradle/
+*.iml
