--- doc/allclasses-frame.html
@@ -1,22 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc (1.8.0_191) on Fri Apr 19 12:56:06 PDT 2019 -->
-<title>All Classes</title>
-<meta name="date" content="2019-04-19">
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<script type="text/javascript" src="script.js"></script>
-</head>
-<body>
-<h1 class="bar">All&nbsp;Classes</h1>
-<div class="indexContainer">
-<ul>
-<li><a href="org/reso/Commander.html" title="class in org.reso" target="classFrame">Commander</a></li>
-<li><a href="org/reso/Commander.Builder.html" title="class in org.reso" target="classFrame">Commander.Builder</a></li>
-<li><a href="org/reso/Main.html" title="class in org.reso" target="classFrame">Main</a></li>
-<li><a href="org/reso/TokenHttpClientFactory.html" title="class in org.reso" target="classFrame">TokenHttpClientFactory</a></li>
-</ul>
-</div>
-</body>
-</html>
