--- doc/index-files/index-2.html
@@ -1,145 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc (1.8.0_191) on Fri Apr 19 12:56:06 PDT 2019 -->
-<title>C-Index</title>
-<meta name="date" content="2019-04-19">
-<link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
-<script type="text/javascript" src="../script.js"></script>
-</head>
-<body>
-<script type="text/javascript"><!--
-    try {
-        if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="C-Index";
-        }
-    }
-    catch(err) {
-    }
-//-->
-</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="topNav"><a name="navbar.top">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
-<a name="navbar.top.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../org/reso/package-summary.html">Package</a></li>
-<li>Class</li>
-<li><a href="../overview-tree.html">Tree</a></li>
-<li><a href="../deprecated-list.html">Deprecated</a></li>
-<li class="navBarCell1Rev">Index</li>
-<li><a href="../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li><a href="index-1.html">Prev Letter</a></li>
-<li><a href="index-3.html">Next Letter</a></li>
-</ul>
-<ul class="navList">
-<li><a href="../index.html?index-files/index-2.html" target="_top">Frames</a></li>
-<li><a href="index-2.html" target="_top">No&nbsp;Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_top">
-<li><a href="../allclasses-noframe.html">All&nbsp;Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_top");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip.navbar.top">
-<!--   -->
-</a></div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<div class="contentContainer"><a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">G</a>&nbsp;<a href="index-4.html">M</a>&nbsp;<a href="index-5.html">O</a>&nbsp;<a href="index-6.html">S</a>&nbsp;<a href="index-7.html">T</a>&nbsp;<a href="index-8.html">U</a>&nbsp;<a href="index-9.html">V</a>&nbsp;<a name="I:C">
-<!--   -->
-</a>
-<h2 class="title">C</h2>
-<dl>
-<dt><span class="memberNameLink"><a href="../org/reso/TokenHttpClientFactory.html#close-org.apache.http.client.HttpClient-">close(HttpClient)</a></span> - Method in class org.reso.<a href="../org/reso/TokenHttpClientFactory.html" title="class in org.reso">TokenHttpClientFactory</a></dt>
-<dd>&nbsp;</dd>
-<dt><a href="../org/reso/Commander.html" title="class in org.reso"><span class="typeNameLink">Commander</span></a> - Class in <a href="../org/reso/package-summary.html">org.reso</a></dt>
-<dd>
-<div class="block">URI
- Most of the work done by the WebAPI commander is done by this class.</div>
-</dd>
-<dt><span class="memberNameLink"><a href="../org/reso/Commander.html#Commander-java.lang.String-java.lang.String-boolean-">Commander(String, String, boolean)</a></span> - Constructor for class org.reso.<a href="../org/reso/Commander.html" title="class in org.reso">Commander</a></dt>
-<dd>
-<div class="block">Creates a Commander instance that uses the given Bearer token for authentication and allows the Client
- to specify whether to use an EdmEnabledClient or normal OData client.</div>
-</dd>
-<dt><a href="../org/reso/Commander.Builder.html" title="class in org.reso"><span class="typeNameLink">Commander.Builder</span></a> - Class in <a href="../org/reso/package-summary.html">org.reso</a></dt>
-<dd>
-<div class="block">Builder pattern implemented for creating Commander instances</div>
-</dd>
-<dt><span class="memberNameLink"><a href="../org/reso/Commander.html#convertMetadata-java.lang.String-">convertMetadata(String)</a></span> - Method in class org.reso.<a href="../org/reso/Commander.html" title="class in org.reso">Commander</a></dt>
-<dd>
-<div class="block">Converts metadata in EDMX format to metadata in Swagger 2.0 format.</div>
-</dd>
-<dt><span class="memberNameLink"><a href="../org/reso/TokenHttpClientFactory.html#create-org.apache.olingo.commons.api.http.HttpMethod-java.net.URI-">create(HttpMethod, URI)</a></span> - Method in class org.reso.<a href="../org/reso/TokenHttpClientFactory.html" title="class in org.reso">TokenHttpClientFactory</a></dt>
-<dd>&nbsp;</dd>
-</dl>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">G</a>&nbsp;<a href="index-4.html">M</a>&nbsp;<a href="index-5.html">O</a>&nbsp;<a href="index-6.html">S</a>&nbsp;<a href="index-7.html">T</a>&nbsp;<a href="index-8.html">U</a>&nbsp;<a href="index-9.html">V</a>&nbsp;</div>
-<!-- ======= START OF BOTTOM NAVBAR ====== -->
-<div class="bottomNav"><a name="navbar.bottom">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
-<a name="navbar.bottom.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../org/reso/package-summary.html">Package</a></li>
-<li>Class</li>
-<li><a href="../overview-tree.html">Tree</a></li>
-<li><a href="../deprecated-list.html">Deprecated</a></li>
-<li class="navBarCell1Rev">Index</li>
-<li><a href="../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li><a href="index-1.html">Prev Letter</a></li>
-<li><a href="index-3.html">Next Letter</a></li>
-</ul>
-<ul class="navList">
-<li><a href="../index.html?index-files/index-2.html" target="_top">Frames</a></li>
-<li><a href="index-2.html" target="_top">No&nbsp;Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_bottom">
-<li><a href="../allclasses-noframe.html">All&nbsp;Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_bottom");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip.navbar.bottom">
-<!--   -->
-</a></div>
-<!-- ======== END OF BOTTOM NAVBAR ======= -->
-</body>
-</html>
