--- doc/package-list
@@ -1,1 +1,0 @@
-org.reso
