--- src/resources/log4j.properties
@@ -1,4 +1,0 @@
-log4j.rootLogger=INFO, Console
-log4j.appender.Console=org.apache.log4j.ConsoleAppender
-log4j.appender.Console.layout=org.apache.log4j.PatternLayout
-log4j.appender.Console.layout.ConversionPattern=[%d] %p: %m%n
