+++ src/test/java/org/reso/commander/MainTest.java
@@ -1,0 +1,13 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package org.reso.commander;
+
+import org.junit.Test;
+
+public class MainTest {
+    @Test public void testAppHasAGreeting() {
+        Main classUnderTest = new Main();
+        //assertNotNull("app should have a greeting", classUnderTest.getGreeting());
+    }
+}
